Teaser view: Defer indexing of the Facet selection field.
Bringing back the Views config update code.  For this to work, we have to defer
setting up indexing on the Facet selection field.  A Directory entry content
type may not have the Facet selection field when the Channel selection field is
added to it.  This breaks the Search index.  Deferring indexing of the Facet
selection field until this field is actually added to a Directory entry content
type resolves this.

Issue #38