Facet group heading: Clearly identify group heading.
Header items should be wrapped in a header tag.  This change was going into the
localgov_theme in the form of an overridden Twig template.  But it looks so
appropriate that we thought it better to include it in the source template.

Issue: https://github.com/localgovdrupal/localgov_theme/issues/85