Facet fixes: corrects required issue where values lost, make required work.
When required the facet field lost values. It should have set the
default to a single facet enabled, if there was only one.

Also required didn't actually stop the submission of the form with
no facets selected.

Fixes and extended tests for different variants. Also adds if there
are no facets available on a channel.