For automated sorting don't include whitespace characters.
"Any chance you can please pass the $sort_value through trim()? I have
found a Directory venue page in our site whose title has a space at the
beginning. Possibly entered by mistake! It is sitting at the top of the
list as a result."

"So I was thinking about this.
My immediate reaction was we should then trim the field itself, rather
than the default value entered into the field, *but* then I thought
if you are entering a string for sorting that's really the value it should
be sorted on with exactly what was entered, so then trim()ing the automated
output starts to make sense, so I'll add that."