Feature/2.x/261 open referral (#136)
Open Referral support for Directories #136 

 * Adds an organisation directory entry type
 * Adds default mappings for Open Referral types to Directory Items and Facets
 * Adds module to expose Venues as Services, and add required Organizations to them.

## Commit history:-

* Functioning notes for /services /services/{id}.

Checking WIP.

A lot of this is moving into the generic localgov_openreferral module.
This module will then include the configuration for mapping and exposing
directory entries to services and organisations (by user configuration
or automatically); and their facets to taxonomy...

No longer relevant, not using rest_views.

Field to referenece organisations.

Add to venues if module is enabled. Still to do if in refresh optional
config if venues enabled after.

Endpoints, normalization etc. moved to localgov_openreferral standalone. Added config to support here.

* Directory type for organisations.

Has no required fields other than title, and can be used as generated on
the fly with entity reference for open referral if it is not being put
into directories itself.

* Script to generate organisations for existing venues #132

* Wakey Wakey. Remove the additional fields used in testing from the default map.

* Smoother venue to Open Referral. #132

* Bugfix: Facet mapping by entity type, bundle. Not entity type, id.

* Spelling error.

https://github.com/localgovdrupal/localgov_directories/pull/136#discussion_r681927682

* Other code, and comment, review recommendations.

* Menu link grouping for new organisation entry type.

https://github.com/localgovdrupal/localgov_directories/pull/136#issuecomment-892773064

* Add Open Referral as suggestion. Should also be used for tests.

https://github.com/localgovdrupal/localgov_directories/pull/136#discussion_r685367904

* Dependency fixes thrown by tests.

* Fix coding standards.

* Fix coding standards.

* Openreferral now has a release.

* Mark Directory Venue Open Referral as experimental.

* Marking _both_ directories OR and venue OR as experimental.

Co-authored-by: Finn <finn@agile.coop>