Fix(Distance): Round figures for distance options.
Solr ignores fraction values in location queries.  So 0.5km is treated as 0km
and 1.5km is treated as 1km.  To avoid confusion, our Location View's filter
now uses round values for distance in Kilometer.  This means the given values
are not fully equivalent to their distance in miles.  For example, we are now
using 3km for 2miles even though the exact value is 3.2km.

It should be noted that Solr's default distance unit is Kilometer whereas
MySQL's is meter.