Add function to fetch sibling facets that would be avalible
This is to address the issue in this comment
  https://github.com/localgovdrupal/localgov_directories/pull/235#issuecomment-1277842772

Instead of just using the facets from the result set, re-run the query, each
time eliminating all the facets except each group, and filter out its own
so there is a list of each sibling facet that could be set with an OR that
also mathces an AND in the other group.
This then produces a filtered list of sibling facets that should be reachable
within the AND accross the groups and OR within them.