Adds test for checking that correct facets are visible during selection
Verifies that when facets are selected, facets that would offer a possible
selection are visible and not hidden. This occurs due to the way facets are
generated from the result set. Since each facet group is an OR, if there
are entries that could be shown by selecting OR in the group whilst
applying the AND filter from the other groups, they need to be selectable.

Adds a test to verify this with comments to help understand what is going on,
as this is a bit on the complicated side. What should happen is that in each
facet group, if after applying an AND filter from the other group, verify
that the correct facets are visible.

Eg. All facets from group 1 which would apply when facet 3 is selected, and all
facets from group 2 which would apply when facet 1 is selected, and then the
results are combined.
In the above case, facet 2 which is only assigned to content with facet 4
would be hidden as it would be inreachable.