Reverse the facet group filter removing the passed in facet filter group
Change the facet filters so that the we test the avalible facets by looping
through each facet group, removing that from the query and seeing what facets
from that group would be returned as an or filter. This is becuase going the
other way (removing the other groups) can lead to situations where possible
facets are removed.