Support ajax views for facet filtering
Since the facet filtering relies on the node route to determine the channel
this breaks when using an ajax view so all facets get shown.
To prevent this, also checks the route name for 'views.ajax' and if on an
ajax view route, use the request stack -> currentRequest() to get the whole
query string and find the view name and argument, using that to load the
channel node.

Note: This introduces
  \Symfony\Component\HttpFoundation\RequestStack
as a new dependency in the create block (should be fine as it is from the
container instead of a service).
Since the views query parameters are sanatized we can't use the request->query
method to retrive them. This therefore splits them from the queryString directly.